+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-
 ## Releases
 
-**0.0.2** Fix validation with empty params.
+### v0.0.2 (2024-11-20)
+
+- Fix validation with empty params.
+- Add more acceptance tests.
+- Improve documentation and examples.